requirements.txt
requests==2.20.0
pyfiglet==0.8.post1
PyMySQL==0.9.3
configparser==3.7.4
spyse.py
tldextract